Generate-code (#18)
* Add pybase62 library

* Add Flask-Alembic library

* Add code column both `participants` and `boards` tables

* Pass `code`, `name` and `score` fields in ParticipantSchema

* Add `code` fields in board schema

* Remove id fields in BoardSchema

* Add `code` column in participant model

* Add `code` column in `board` model

* Re-work sql query get to board by `code`

* Assing `code` variable for new_board with encode method when called `POST` endpoint insede of create_board method

* Re-work sql query get board by `code` parameter where come from uri parameter

* Re-name `origins`' value parameter `CORS_ORIGINS` instead of 'ORIGINS'

* Import base62 library

* Re-name `origins`' value parameter `CORS_ORIGINS` instead of 'ORIGINS'

* Re-name `origins`' value parameter `CORS_ORIGINS` instead of 'ORIGINS'

* Import base62 library

* Import Board class from board model

* Import Board class from board model

* Re-work add_score_by_id method and sql query to use `participant.code` and `board_code` parameter instead of `board .id` and `participant.id`. Catch the invalid board and participant `code` value

* Re-work create_participants method and sql query to use `board.code` instead of `board.id` and catch if exists `board.id` in `board` table

* Re-work list_participants to list by `code` parameter with `board_code` instead of `board_id`

* Re-work remove_participants_by_id method to delete particiapants with `participant`'s code and `board_code` parameters

* Re-work modify_participants_by_id method to modify participant's `name` variable by `participants`'s `code` parameter and `board_code` parameter. Catch if `board.id` already exists

* ValueError exception

* Update alembic version files

* Add nullable, unique and index constraints for `code` column

* Add nullable, unique and index constraints for `board table`'s `code` column

* Create `handle_not_found` method in error_handlers.py

* Add nullable, unique and index constraints both `participants` and `boards` tables

* Remove `boards_schema`

* Import exception and error_handlers modules

* Register error_handlers for app

* Create handle_exception method to catch HTTPExceptions

* Remove boards_schema method from `schema/board`

* Import error handlers and exception modules

* Re-name `get_board_by_id` to `get_board_by_code`

* Raise NotFound exception if board is None

* Re-name method `modify_board_by_id` to `modify_board_by_code`

* Raise NotFound exception if board is None

* Fetch and execute next_board _id value and pass the encode method

* Import NotFound exception

* Re-name  method `add_score_by_id` to `add_score_by_code`

* Re-name `is_valid` variable to `board`

* Check board if exist add score by participant code and increment value and return participant

* Publish channel by `participant.code` parameter

* Import Sequence query method from sqlalchemy

* Import NotFound exception

* Remove try-except lines

* Execute sql query if board is not None and catch NotFound exception if board is None

* Create new participant and set `code` column with `base62.encode` method by next_participant_id value

* Fetch next_participant_id value

* Get board by board_code if the board is None raise NotFound() method if it's not fetch and return participant

* Re-name `remove_participants_by_id` to `remove_participants_by_code

* Remove try-except lines

* Re-name `is_valid` to `board`

* Remove `is_valid` variable

* Raise NotFound() if board is None

* Delete participant by code and board_code and return '204' if successful

* Re-name `modify_participants_by_id` to `modify_participants_by_code`

* Remove try-except lines

* Re-name `is_valid` to `board`

* Check board is None or exist if it's None raise NotFound() error

* If board and participant are exist  modify the participant and return the participant

* Update schema/board.py use Meta class for specific fields and nested fields

* Use `SQLAlchemySchema` from marshmallow library

* Define columns manuel instead of using fields

* Define columns manuel instead of using fields

* Use `SQLAlchemySchema` instead of `SQLAlchemyAutoSchema`

* Remove class Meta

* Define fields manuel

* Define fields manuel

* Remove class Meta

+++ app/api/participant.py
@@ -1,47 +1,81 @@
 from flask import Blueprint, request, jsonify
+from sqlalchemy import Sequence
+from werkzeug.exceptions import NotFound
 from app import db, app
+from app.models.board import Board
 from app.models.participant import Participant
 from app.schema.participant import participant_schema, participants_schema
 from flask_cors import CORS
+import base62
 
-participants = Blueprint('participants', __name__, url_prefix='/api/v1/boards/<board_id>/participants')
-CORS(participants, resources={r"/api/*": {"origins": app.config.get('ORIGINS')}})
+participants = Blueprint('participants', __name__, url_prefix='/api/v1/boards/<board_code>/participants')
+CORS(participants, resources={r"/api/*": {"origins": app.config.get('CORS_ORIGINS')}})
 
 @participants.route('', methods=["POST"])
-def create_participants(board_id):
-    name = request.json['name']
+def create_participants(board_code):
+    board = db.session.query(Board)\
+        .filter(Board.code == board_code)\
+        .first()
 
-    participant = Participant(name=name, board_id=board_id)
+    if board is None:
+        raise NotFound()
+
+    name = request.json['name']
+    participant = Participant(name=name, board_id=board.id)
+    participant.id = db.session.execute(Sequence("participants_id_seq"))
+    participant.code = base62.encode(hash(('participants', participant.id)), 8)[-8:]
 
     db.session.add(participant)
     db.session.commit()
+
     return participant_schema.dump(participant)
 
 @participants.route('', methods=["GET"])
-def list_participants(board_id):
-    participants = db.session.query(Participant)\
-        .filter(Participant.board_id == board_id)\
+def list_participants(board_code):
+    board = db.session.query(Board) \
+        .filter(Board.code == board_code) \
+        .first()
+
+    if board is None:
+        raise NotFound()
+
+    participant_list = db.session.query(Participant) \
+        .filter(Participant.board_id == board.id) \
         .all()
 
-    return jsonify(participants_schema.dump(participants))
+    return jsonify(participants_schema.dump(participant_list))
 
-@participants.route('/<id>', methods=["DELETE"])
-def remove_participants_by_id(id, board_id):
-    participant = db.session.query(Participant)\
-        .filter(Participant.board_id == board_id)\
-        .filter(Participant.id == id)\
+@participants.route('/<code>', methods=["DELETE"])
+def remove_participants_by_code(code, board_code):
+    board = db.session.query(Board)\
+    .filter(Board.code == board_code)\
+    .first()
+
+    if board is None:
+        raise NotFound()
+
+    participant = db.session.query(Participant) \
+        .filter(Participant.board_id == board.id) \
+        .filter(Participant.code == code) \
         .first()
 
     db.session.delete(participant)
     db.session.commit()
 
-    return ('', 204)
+    return '', 204
 
-@participants.route('/<id>/name', methods=["PUT"])
-def modify_participants_by_id(id, board_id):
-    participant = db.session.query(Participant)\
-        .filter(Participant.board_id == board_id)\
-        .filter(Participant.id == id)\
+@participants.route('/<code>/name', methods=["PUT"])
+def modify_participants_by_code(code, board_code):
+    board = db.session.query(Board)\
+    .filter(Board.code == board_code)\
+    .first()
+
+    if board is None:
+        raise NotFound()
+
+    participant = db.session.query(Participant) \
+        .filter(Participant.board_id == board.id) \
+        .filter(Participant.code == code) \
         .first()
 
     name = request.json['name']
@@ -50,3 +84,4 @@ def modify_participants_by_id(id, board_id):
     db.session.commit()
 
     return jsonify(participant_schema.dump(participant))
+
